+++ src/test/java/com/global/api/tests/network/nts/NTSPOSConfigurationMessageTest.java
@@ -74,7 +74,7 @@ public class NTSPOSConfigurationMessageTest {
             config = new NetworkGatewayConfig(Target.NTS);
             config.setPrimaryEndpoint("test.txns-c.secureexchange.net");
             config.setPrimaryPort(15031);
-            config.setSecondaryEndpoint("test.txns.secureexchange.net");
+            config.setSecondaryEndpoint("test.txns-e.secureexchange.net");
             config.setSecondaryPort(15031);
             config.setEnableLogging(true);
             config.setStanProvider(StanGenerator.getInstance());